Fixes Credential Provider & documentation
