+++ Source/AwsCommonRuntimeKit/auth/credentials/CRTCredentialsProvider.swift
@@ -6,5 +6,4 @@ import AwsCAuth
 public protocol CRTCredentialsProvider {
     var allocator: Allocator {get set}
     func getCredentials() async throws -> CRTCredentials
-
 }
